# wear, spalling, and gamepad