… img/spalling.webp
… img/wear.webp
… index.html
… todo.txt