tech: wear - 2x harpoon fire rate, 10% chance for rope to snap after hitting a mob (losing ammo)
tech: spalling - after harpoon breaks get 2 research and 1 boost

perfect diamagnetism field float also accelerates horizontally constraint balance: after 40->50 seconds spawn WIMPs  (then spawn one every 7->9 seconds) constraint balance: 0.5x max health now no longer considers difficulty based heal reduction when returning lost health plasma dielectric 3->2 cost
WIMPs 6->7 research per level
mine tech sentry fires ~1.2x faster
mass-energy equivalence 1.3 ->1.5 damage

added button in settings to load gamepad code: https://github.com/kgurchiek/n-gon-controller

fixed bug with wormhole crashing after eating very small power ups
Mod info: The mod option for a controller is still helpful, as it is automatic versus requiring a button every time.