@btcool 2025-04-14
Yeah, imma just do it manually.